backend/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-data-tables
azure-ai-formrecognizer
pandas
thefuzz
python-Levenshtein
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dropbox
